requirements.txt
transformers==4.30.2
torch==2.0.1
numpy==1.24.3
opencv-python==4.8.0.74
scikit-learn==1.2.2
pandas==2.0.2
fastapi==0.100.0
uvicorn==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.11.2
pyttsx3==2.90
python-telegram-bot==20.6
googletrans==3.1.0a0
pytesseract==0.3.10
pillow==10.0.0
sounddevice==0.4.6
scipy==1.10.1
networkx==3.1
sympy==1.12
spacy==3.5.3
face_recognition==1.3.0
sklearn==1.2.2
nltk==3.8.1
sentence-transformers==2.2.2
openai==0.27.8
pyaudio==0.2.13
mediapipe==0.10.0
pydub==0.25.1
moviepy==1.0.3
pyautogui==0.9.54
pynput==1.7.6
jinja2==3.1.2
aiofiles==23.1.0
